Refactor more DeleteSourceDialog to `gui.source` namespace
